Move file system runtime modules to types
