+++ front/lib/api/actions/servers/user_memory/tools/index.ts
@@ -6,7 +6,6 @@ import {
   USER_MEMORY_READ_TOOL_NAME,
   USER_MEMORY_TOOLS_METADATA,
 } from "@app/lib/api/actions/servers/user_memory/metadata";
-import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
 import { getUpdatedContentAndOccurrences } from "@app/lib/api/files/utils";
 import {
   exceedsUserMemoryLimit,
@@ -15,6 +14,7 @@ import {
 } from "@app/lib/api/user_memory";
 import type { Authenticator } from "@app/lib/auth";
 import { MAX_USER_MEMORY_CONTENT_LENGTH } from "@app/types/api/me/memory";
+import { DustFileSystem } from "@app/types/dust_file_system";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
