+++ front/lib/api/file_system/index.ts
@@ -1,10 +1,10 @@
+export type { FileSystemEntry } from "@app/types/api/file_system/types";
 export {
   DustFileSystem,
   DustFileSystemError,
   parseScopedPrefix,
   sanitizeFileSystemName,
-} from "@app/lib/api/file_system/dust_file_system";
-export type { FileSystemEntry } from "@app/types/api/file_system/types";
+} from "@app/types/dust_file_system";
 export type {
   DustFileSystemErrorCode,
   FileSystemMount,
