+++ front/lib/api/file_system/upload_from_url.test.ts
@@ -1,8 +1,8 @@
 import { createConversation } from "@app/lib/api/assistant/conversation";
-import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
 import { uploadFileFromUrlToFileSystem } from "@app/lib/api/file_system/upload_from_url";
 import { untrustedFetch } from "@app/lib/egress/server";
 import { createResourceTest } from "@app/tests/utils/generic_resource_tests";
+import { DustFileSystem } from "@app/types/dust_file_system";
 import { isString } from "@app/types/shared/utils/general";
 import { Readable } from "stream";
 import { describe, expect, it, vi } from "vitest";
