+++ front/lib/api/file_system/upload_from_url.ts
@@ -2,9 +2,9 @@ import {
   frameFileCreateRejectedError,
   frameFileEditRejectedError,
 } from "@app/lib/api/actions/servers/files/tools/utils";
-import type { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
 import { validateExternalUrl } from "@app/lib/api/url_safety";
 import { untrustedFetch } from "@app/lib/egress/server";
+import type { DustFileSystem } from "@app/types/dust_file_system";
 import {
   getFileFormatCategory,
   isInteractiveContentType,
