+++ front/lib/api/files/mount_path.test.ts
@@ -1,3 +1,5 @@
+import { FileFactory } from "@app/tests/utils/FileFactory";
+import { createResourceTest } from "@app/tests/utils/generic_resource_tests";
 import {
   disambiguateFileName,
   getBaseMountPathForWorkspace,
@@ -24,9 +26,7 @@ import {
   resolveScopedMountFilePath,
   splitFrameEntryScopedPath,
   validateMountFolderName,
-} from "@app/lib/api/files/mount_path";
-import { FileFactory } from "@app/tests/utils/FileFactory";
-import { createResourceTest } from "@app/tests/utils/generic_resource_tests";
+} from "@app/types/mount_path";
 import { describe, expect, it } from "vitest";
 
 describe("mount_path helpers", () => {
