+++ front/lib/api/sandbox_functions/dsbx_db.ts
@@ -1,11 +1,5 @@
 import { randomUUID } from "node:crypto";
 import path from "node:path";
-import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
-import {
-  POD_SANDBOX_DATABASES_DIR,
-  podDatabaseExecEnvVars,
-  TOOL_OUTPUTS_FOLDER_NAME,
-} from "@app/lib/api/files/mount_path";
 import { getRedisStreamClient } from "@app/lib/api/redis";
 import { isValidPodDatabaseName } from "@app/lib/api/sandbox/db";
 import { ensurePodSandboxReady } from "@app/lib/api/sandbox/lifecycle";
@@ -28,7 +22,13 @@ import type { SandboxResource } from "@app/lib/resources/sandbox_resource";
 import type { SpaceResource } from "@app/lib/resources/space_resource";
 import logger from "@app/logger/logger";
 import tracer from "@app/logger/tracer";
+import { DustFileSystem } from "@app/types/dust_file_system";
 import { SCOPED_PREFIX_POD } from "@app/types/file_system";
+import {
+  POD_SANDBOX_DATABASES_DIR,
+  podDatabaseExecEnvVars,
+  TOOL_OUTPUTS_FOLDER_NAME,
+} from "@app/types/mount_path";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
