+++ front/lib/api/sandbox_functions/publish_sandbox_function.ts
@@ -1,4 +1,3 @@
-import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
 import { buildSandboxFunctionOnSandbox } from "@app/lib/api/sandbox_functions/build_on_sandbox";
 import { SandboxFunctionError } from "@app/lib/api/sandbox_functions/errors";
 import { deriveSandboxFunctionSlug } from "@app/lib/api/sandbox_functions/slug";
@@ -10,6 +9,7 @@ import {
 } from "@app/lib/resources/sandbox_function_resource";
 import type { SpaceResource } from "@app/lib/resources/space_resource";
 import type { SandboxFunctionExecutionMode } from "@app/types/api/sandbox_functions";
+import { DustFileSystem } from "@app/types/dust_file_system";
 import { sandboxFunctionContentType } from "@app/types/files";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
