+++ front/lib/api/skills/conversation_files.ts
@@ -1,7 +1,7 @@
-import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
 import type { Authenticator } from "@app/lib/auth";
 import type { SkillResource } from "@app/lib/resources/skill/skill_resource";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
+import { DustFileSystem } from "@app/types/dust_file_system";
 import type { Result } from "@app/types/shared/result";
 import { Ok } from "@app/types/shared/result";
 import type { Readable } from "stream";
