+++ front/lib/api/user_memory.ts
@@ -1,7 +1,7 @@
-import type { DustFileSystemError } from "@app/lib/api/file_system/dust_file_system";
-import { DustFileSystem } from "@app/lib/api/file_system/dust_file_system";
 import type { Authenticator } from "@app/lib/auth";
 import { MAX_USER_MEMORY_CONTENT_LENGTH } from "@app/types/api/me/memory";
+import type { DustFileSystemError } from "@app/types/dust_file_system";
+import { DustFileSystem } from "@app/types/dust_file_system";
 import { userScopedPath } from "@app/types/file_system";
 import type { Result } from "@app/types/shared/result";
 import { Ok } from "@app/types/shared/result";
