+++ front/lib/api/viz/extract_file_refs.ts
@@ -1,7 +1,8 @@
 // Ported from viz/app/lib/parseFileRefs.ts — keep classification in sync with
-// `isAgentScopedPath` in front/lib/api/files/mount_path.ts until deduped into a shared package.
-import { isAgentScopedPath } from "@app/lib/api/files/mount_path";
+// `isAgentScopedPath` in front/types/mount_path.ts until deduped into a shared package.
+
 import logger from "@app/logger/logger";
+import { isAgentScopedPath } from "@app/types/mount_path";
 import ts from "typescript";
 
 export type FileRef =
